Set hypertexnames to false
Adding hypertexnames=false to \hypersetup fixes the links in the table of contents so they work properly even if there is front matter. Before this change, clicking on the first section of the main report jumped the document to the first front matter section. Now the links work for me as expected.